refactor: move Layout into SharedComponents
