+++ src/components/ErrorView/ErrorView.tsx
@@ -2,9 +2,12 @@ import { FC } from "react";
 import styled from "styled-components";
 
 import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
-import Layout from "../Layout";
-import { HeaderVariant, FooterVariant } from "../Layout/Layout";
+import {
+  HeaderVariant,
+  FooterVariant,
+} from "../SharedComponents/Layout/Layout";
 
+import Layout from "@/components/SharedComponents/Layout";
 import { P, Heading } from "@/components/SharedComponents/Typography";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import ButtonGroup from "@/components/SharedComponents/ButtonGroup";
