+++ src/components/SharedComponents/Layout/Layout.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 
-import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
-import renderWithProviders from "../../__tests__/__helpers__/renderWithProviders";
+import { DEFAULT_SEO_PROPS } from "../../../browser-lib/seo/Seo";
+import renderWithProviders from "../../../__tests__/__helpers__/renderWithProviders";
 
 import Layout from "./Layout";
